+++ requirements.txt
@@ -1,0 +1,146 @@
+aiohttp==3.9.5
+aiosignal==1.3.1
+alembic==1.13.1
+annotated-types==0.6.0
+anthropic==0.25.6
+anyio==4.3.0
+appdirs==1.4.4
+asgiref==3.8.1
+async-timeout==4.0.3
+attrs==23.2.0
+backoff==2.2.1
+bcrypt==4.1.2
+beautifulsoup4==4.12.3
+blinker==1.7.0
+build==1.2.1
+cachetools==5.3.3
+certifi==2024.2.2
+charset-normalizer==3.3.2
+chroma-hnswlib==0.7.3
+chromadb==0.4.24
+click==8.1.7
+coloredlogs==15.0.1
+contextlib2==21.6.0
+crewai==0.28.8
+dataclasses-json==0.6.4
+Deprecated==1.2.14
+distro==1.9.0
+docstring-parser==0.15
+embedchain==0.1.102
+exceptiongroup==1.2.1
+fastapi==0.110.2
+filelock==3.13.4
+Flask==3.0.3
+flatbuffers==24.3.25
+frozenlist==1.4.1
+fsspec==2024.3.1
+google-api-core==2.18.0
+google-auth==2.29.0
+google-cloud-aiplatform==1.48.0
+google-cloud-bigquery==3.21.0
+google-cloud-core==2.4.1
+google-cloud-resource-manager==1.12.3
+google-cloud-storage==2.16.0
+google-crc32c==1.5.0
+google-resumable-media==2.7.0
+googleapis-common-protos==1.63.0
+gptcache==0.1.43
+grpc-google-iam-v1==0.13.0
+grpcio==1.62.2
+grpcio-status==1.62.2
+h11==0.14.0
+httpcore==1.0.5
+httptools==0.6.1
+httpx==0.27.0
+huggingface-hub==0.22.2
+humanfriendly==10.0
+idna==3.7
+importlib-metadata==7.0.0
+importlib_resources==6.4.0
+instructor==0.5.2
+itsdangerous==2.2.0
+Jinja2==3.1.3
+jsonpatch==1.33
+jsonpointer==2.4
+kubernetes==29.0.0
+langchain==0.1.16
+langchain-community==0.0.34
+langchain-core==0.1.45
+langchain-openai==0.0.5
+langchain-text-splitters==0.0.1
+langsmith==0.1.50
+Mako==1.3.3
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+marshmallow==3.21.1
+mdurl==0.1.2
+mmh3==4.1.0
+monotonic==1.6
+mpmath==1.3.0
+multidict==6.0.5
+mypy-extensions==1.0.0
+numpy==1.26.4
+oauthlib==3.2.2
+onnxruntime==1.17.3
+openai==1.23.4
+opentelemetry-api==1.24.0
+opentelemetry-exporter-otlp-proto-common==1.24.0
+opentelemetry-exporter-otlp-proto-grpc==1.24.0
+opentelemetry-exporter-otlp-proto-http==1.24.0
+opentelemetry-instrumentation==0.45b0
+opentelemetry-instrumentation-asgi==0.45b0
+opentelemetry-instrumentation-fastapi==0.45b0
+opentelemetry-proto==1.24.0
+opentelemetry-sdk==1.24.0
+opentelemetry-semantic-conventions==0.45b0
+opentelemetry-util-http==0.45b0
+orjson==3.10.1
+overrides==7.7.0
+packaging==23.2
+posthog==3.5.0
+proto-plus==1.23.0
+protobuf==4.25.3
+pulsar-client==3.5.0
+pyasn1==0.6.0
+pyasn1_modules==0.4.0
+pydantic==2.7.1
+pydantic_core==2.18.2
+Pygments==2.17.2
+pypdf==3.17.4
+PyPika==0.48.9
+pyproject_hooks==1.0.0
+pysbd==0.3.4
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.0
+PyYAML==6.0.1
+regex==2023.12.25
+requests==2.31.0
+requests-oauthlib==2.0.0
+rich==13.7.1
+rsa==4.9
+schema==0.7.5
+shapely==2.0.4
+six==1.16.0
+sniffio==1.3.1
+soupsieve==2.5
+SQLAlchemy==2.0.29
+starlette==0.37.2
+sympy==1.12
+tenacity==8.2.3
+tiktoken==0.5.2
+tokenizers==0.19.1
+tomli==2.0.1
+tqdm==4.66.2
+typer==0.9.4
+typing-inspect==0.9.0
+typing_extensions==4.11.0
+urllib3==2.2.1
+uvicorn==0.29.0
+uvloop==0.19.0
+watchfiles==0.21.0
+websocket-client==1.8.0
+websockets==12.0
+Werkzeug==3.0.2
+wrapt==1.16.0
+yarl==1.9.4
+zipp==3.18.1
